Refactor Model test hierarchy to have ChartDataModel2DTest class
